requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
pydantic[email]
sqlmodel
fastapi-cache2[redis]
python-multipart
psycopg2